FCOM-331: Added fake SMTP for unit tests + minor code and document changes

+++ Frends.Community.Email.Tests/SendEmailTests.cs
@@ -1,27 +1,28 @@
 ﻿using NUnit.Framework;
 using System.IO;
+using System.Threading;
 
 namespace Frends.Community.Email.Tests
-{   
-     /// <summary>
-     /// NOTE: To run these unit tests, you need an SMTP test server. Fill in the properties below with your values.
-     /// </summary>
+{
+    /// <summary>
+    /// NOTE: To run these unit tests, you need an SMTP test server.
+    /// 
+    /// docker run -p 3000:80 -p 2525:25 -d rnwood/smtp4dev:v3
+    /// 
+    /// Management console can be found from http://localhost:3000/
+    /// 
+    /// </summary>
     [TestFixture]
-    [Ignore("Tests requires external SMTP server to work")]
     public class SendEmailTests
     {
-        // ****************************************** FILL THESE ******************************************************
-        private const string USERNAME = "";
-        private const string PASSWORD = "";
-        private const string SMTPADDRESS = "";
-        private const string TOEMAILADDRESS = "";
-        private const string FROMEMAILADDRESS = "";
-        private const int PORT = 465;
-        private const bool USESSL = true;
-        private const bool ACCEPTALLCERTS = false;
-        // ************************************************************************************************************
-
-
+        private const string USERNAME = "test";
+        private const string PASSWORD = "test";
+        private const string SMTPADDRESS = "localhost";
+        private const string TOEMAILADDRESS = "test@test.com";
+        private const string FROMEMAILADDRESS = "user@user.com";
+        private const int PORT = 2525;
+        private const bool USESSL = false;
+        private const bool ACCEPTALLCERTS = true;
         private const string TEMP_ATTACHMENT_SOURCE = "emailtestattachments";
         private const string TEST_FILE_NAME = "testattachment.txt";
         private const string TEST_FILE_NOT_EXISTING = "doesntexist.txt";
@@ -36,13 +37,11 @@ namespace Frends.Community.Email.Tests
         {
             _localAttachmentFolder = Path.Combine(Path.GetTempPath(), TEMP_ATTACHMENT_SOURCE);
 
-            if (!Directory.Exists(_localAttachmentFolder))
-                Directory.CreateDirectory(_localAttachmentFolder);
+            if (!Directory.Exists(_localAttachmentFolder)) Directory.CreateDirectory(_localAttachmentFolder);
 
             _filepath = Path.Combine(_localAttachmentFolder, TEST_FILE_NAME);
 
-            if (!File.Exists(_filepath))
-                using (File.Create(_filepath)) { }
+            if (!File.Exists(_filepath)) using (File.Create(_filepath)) { }
 
             _input = new Input()
             {
@@ -70,8 +69,7 @@ namespace Frends.Community.Email.Tests
         [TearDown]
         public void EmailTestTearDown()
         {
-            if (Directory.Exists(_localAttachmentFolder))
-                Directory.Delete(_localAttachmentFolder, true);
+            if (Directory.Exists(_localAttachmentFolder)) Directory.Delete(_localAttachmentFolder, true);
         }
 
         [Test]
@@ -80,7 +78,7 @@ namespace Frends.Community.Email.Tests
             var input = _input;
             input.Subject = "Email test - PlainText";
             
-            var result = EmailTask.SendEmail(input, null, _options, new System.Threading.CancellationToken());
+            var result = EmailTask.SendEmail(input, null, _options, new CancellationToken());
             Assert.IsTrue(result.EmailSent);
         }
 
@@ -100,7 +98,7 @@ namespace Frends.Community.Email.Tests
 
             var Attachments = new Attachment[] { attachment };
 
-            var result = EmailTask.SendEmail(input, Attachments, _options, new System.Threading.CancellationToken());
+            var result = EmailTask.SendEmail(input, Attachments, _options, new CancellationToken());
             Assert.IsTrue(result.EmailSent);
         }
 
@@ -117,7 +115,7 @@ namespace Frends.Community.Email.Tests
             };
             var Attachments = new Attachment[] { attachment};
 
-            var result = EmailTask.SendEmail(input, Attachments, _options, new System.Threading.CancellationToken());
+            var result = EmailTask.SendEmail(input, Attachments, _options, new CancellationToken());
             Assert.IsTrue(result.EmailSent);
         }
 
@@ -137,7 +135,7 @@ namespace Frends.Community.Email.Tests
 
             var Attachments = new Attachment[] { attachment };
 
-            var result = EmailTask.SendEmail(input, Attachments, _options, new System.Threading.CancellationToken());
+            var result = EmailTask.SendEmail(input, Attachments, _options, new CancellationToken());
             Assert.IsFalse(result.EmailSent);
         }
 
@@ -157,7 +155,7 @@ namespace Frends.Community.Email.Tests
 
             var Attachments = new Attachment[] { attachment };
 
-            Assert.Throws<FileNotFoundException>(() => EmailTask.SendEmail(input, Attachments, _options, new System.Threading.CancellationToken()));
+            Assert.Throws<FileNotFoundException>(() => EmailTask.SendEmail(input, Attachments, _options, new CancellationToken()));
             
         }
     }
